# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Вовремя — API календаря и бота MAX",
    "version": "1.1.0",
    "description": "Личные данные доступны только владельцу сессии. Production: подписанные initData MAX. Демо: отдельный гостевой профиль. Публичный адрес из APP_URL необходимо проверить перед сдачей."
  },
  "servers": [
    {
      "url": "http://localhost:3002/api",
      "description": "Локальное Docker-демо"
    },
    {
      "url": "https://84.237.53.134/api",
      "description": "Адрес из текущей конфигурации; внешняя доступность требует проверки"
    }
  ],
  "paths": {
    "/config": {
      "get": {
        "summary": "Режим работы и ссылка на бота",
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["demo", "botUrl"],
                  "properties": {
                    "demo": { "type": "boolean" },
                    "botUrl": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Готовность HTTP-сервера (не гарантия доставки MAX)",
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["ok"],
                  "properties": { "ok": { "type": "boolean", "enum": [true] } }
                }
              }
            }
          }
        }
      }
    },
    "/session": {
      "post": {
        "summary": "Вход с проверенной подписью MAX; пустое тело разрешено только в демо",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "initData": { "type": "string", "maxLength": 20000 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["token", "user"],
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "security": [{ "session": [] }],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        },
        "summary": "Текущий пользователь"
      }
    },
    "/events": {
      "get": {
        "security": [{ "session": [] }],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Event" }
                }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        },
        "summary": "События текущего пользователя по дате"
      },
      "post": {
        "security": [{ "session": [] }],
        "responses": {
          "201": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        },
        "summary": "Создать событие; максимум 500",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EventInput" }
            }
          }
        }
      }
    },
    "/events/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" }
        }
      ],
      "put": {
        "security": [{ "session": [] }],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Событие не принадлежит пользователю или не найдено",
            "content": { "text/plain": { "schema": { "type": "string" } } }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        },
        "summary": "Изменить событие и пересчитать срок",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EventInput" }
            }
          }
        }
      },
      "patch": {
        "security": [{ "session": [] }],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": { "text/plain": { "schema": { "type": "string" } } }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        },
        "summary": "Завершить или вернуть событие",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["completed"],
                "properties": { "completed": { "type": "boolean" } }
              }
            }
          }
        }
      },
      "delete": {
        "security": [{ "session": [] }],
        "responses": {
          "204": { "description": "Удалено" },
          "401": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Не найдено",
            "content": { "text/plain": { "schema": { "type": "string" } } }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        },
        "summary": "Удалить событие с историей доставок"
      }
    },
    "/settings": {
      "put": {
        "security": [{ "session": [] }],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Settings" }
              }
            }
          },
          "400": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        },
        "summary": "Время, часовой пояс и согласие на сообщения",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Settings" }
            }
          }
        }
      }
    },
    "/deliveries": {
      "get": {
        "security": [{ "session": [] }],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Delivery" }
                }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        },
        "summary": "Последние 100 доставок текущего пользователя"
      }
    },
    "/demo/seed": {
      "post": {
        "security": [{ "session": [] }],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Event" }
                }
              }
            }
          },
          "401": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Недоступно для реального профиля",
            "content": { "text/plain": { "schema": { "type": "string" } } }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        },
        "summary": "Синтетические примеры; только пустой демопрофиль"
      }
    },
    "/data": {
      "delete": {
        "security": [{ "session": [] }],
        "responses": {
          "204": { "description": "Удалено" },
          "401": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        },
        "summary": "Удалить события и их доставки, выключить напоминания; профиль остаётся"
      }
    },
    "/max/webhook": {
      "post": {
        "summary": "Входящие обновления MAX, постоянная очередь с подавлением повторов",
        "security": [{ "webhook": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["update_type", "timestamp"],
                "properties": {
                  "update_type": {
                    "type": "string",
                    "enum": [
                      "bot_started",
                      "bot_stopped",
                      "message_created",
                      "message_callback"
                    ]
                  },
                  "timestamp": { "type": "number" },
                  "user": { "type": "object", "additionalProperties": true },
                  "message": { "type": "object", "additionalProperties": true },
                  "callback": {
                    "type": "object",
                    "required": ["callback_id", "payload", "user"],
                    "properties": {
                      "callback_id": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 256
                      },
                      "payload": { "type": "string", "maxLength": 128 },
                      "user": { "type": "object", "additionalProperties": true }
                    }
                  }
                },
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Принято или неподдерживаемый тип пропущен",
            "content": { "text/plain": { "schema": { "type": "string" } } }
          },
          "400": {
            "description": "Не указан timestamp",
            "content": { "text/plain": { "schema": { "type": "string" } } }
          },
          "403": {
            "description": "Неверный секрет",
            "content": { "text/plain": { "schema": { "type": "string" } } }
          },
          "429": {
            "description": "Ошибка запроса",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "http",
        "scheme": "bearer",
        "description": "Токен из POST /session"
      },
      "webhook": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Max-Bot-Api-Secret"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": { "error": { "type": "string" } }
      },
      "EventInput": {
        "type": "object",
        "required": ["templateId", "title", "category", "baseDate"],
        "properties": {
          "templateId": {
            "type": "string",
            "enum": [
              "passport",
              "fluorography",
              "tax",
              "international",
              "insurance",
              "rf-004",
              "rf-009",
              "rf-010",
              "rf-012",
              "rf-015",
              "rf-018",
              "rf-020",
              "rf-029",
              "rf-030",
              "rf-031",
              "rf-035",
              "rf-047",
              "rf-048",
              "rf-050",
              "rf-051",
              "rf-053",
              "rf-054",
              "rf-062",
              "rf-071",
              "rf-074",
              "rf-075",
              "rf-085",
              "rf-087",
              "rf-092",
              "rf-093",
              "rf-094",
              "rf-103",
              "rf-108",
              "rf-112",
              "rf-116",
              "rf-120",
              "rf-122",
              "rf-124",
              "rf-125",
              "rf-126",
              "rf-128",
              "rf-129",
              "rf-147",
              "rf-150",
              "rf-152",
              "custom"
            ]
          },
          "title": { "type": "string", "minLength": 1, "maxLength": 120 },
          "category": {
            "type": "string",
            "enum": [
              "documents",
              "transport",
              "payments",
              "home",
              "health",
              "social",
              "family",
              "education",
              "work",
              "migration",
              "other"
            ]
          },
          "baseDate": {
            "type": "string",
            "format": "date",
            "description": "Существующая календарная дата, 1900–2200"
          },
          "age": {
            "type": "integer",
            "enum": [20, 45],
            "description": "Обязателен для passport"
          },
          "intervalMonths": {
            "type": "integer",
            "minimum": 1,
            "maximum": 120,
            "description": "Обязателен для fluorography; указывает пользователь по рекомендации врача"
          },
          "reminders": {
            "type": "array",
            "maxItems": 8,
            "items": { "type": "integer", "minimum": 0, "maximum": 365 },
            "default": [30, 7, 1, 0]
          },
          "notes": { "type": "string", "maxLength": 2000, "default": "" },
          "documentName": { "type": "string", "maxLength": 160, "default": "" }
        }
      },
      "Event": {
        "allOf": [
          { "$ref": "#/components/schemas/EventInput" },
          {
            "type": "object",
            "required": ["id", "dueDate", "completed", "createdAt"],
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "dueDate": {
                "type": "string",
                "format": "date",
                "description": "Существующая календарная дата, 1900–2200"
              },
              "milestoneDate": {
                "type": "string",
                "format": "date",
                "description": "Существующая календарная дата, 1900–2200"
              },
              "completed": { "type": "boolean" },
              "createdAt": { "type": "string", "format": "date-time" }
            }
          }
        ]
      },
      "Settings": {
        "type": "object",
        "required": ["enabled", "hour", "timezone", "privateMessages"],
        "properties": {
          "enabled": { "type": "boolean" },
          "hour": { "type": "integer", "minimum": 0, "maximum": 23 },
          "timezone": { "type": "string", "example": "Asia/Novosibirsk" },
          "privateMessages": { "type": "boolean" }
        }
      },
      "User": {
        "type": "object",
        "required": ["id", "name", "settings", "botStarted", "demo"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "settings": { "$ref": "#/components/schemas/Settings" },
          "botStarted": { "type": "boolean" },
          "demo": { "type": "boolean" }
        }
      },
      "Delivery": {
        "type": "object",
        "required": ["id", "eventId", "title", "status", "sentAt"],
        "properties": {
          "id": { "type": "string" },
          "eventId": { "type": "string" },
          "title": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["pending", "sending", "sent", "failed"]
          },
          "sentAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      }
    }
  }
}